refactor: separate workflow loading and expressions
