+++ src/workflows/file.ts
@@ -1,6 +1,21 @@
+import { loadWorkflowFile, isApprovalStep, isInputStep } from "./load.js";
+import {
+	resolveArgsTemplate,
+	resolveTemplate,
+	parseStepRef,
+	getStepRefValue,
+	evaluateCondition,
+} from "./expressions.js";
+import type {
+	WorkflowFile,
+	ParallelBranch,
+	WorkflowStep,
+	WorkflowApprovalIdentity,
+	WorkflowStepResult,
+	WorkflowRunResult,
+} from "./types.js";
 import { promises as fsp } from "node:fs";
 import path from "node:path";
-import { parse as parseYaml } from "yaml";
 
 import { randomUUID } from "node:crypto";
 import { isDeepStrictEqual } from "node:util";
@@ -35,7 +50,7 @@ import type {
 	LlmSpendLedger,
 } from "../commands/stdlib/llm_invoke.js";
 import { CostLimitExceededError, CostTracker } from "../core/cost_tracker.js";
-import type { CostLimit, CostSummary } from "../core/cost_tracker.js";
+import type { CostSummary } from "../core/cost_tracker.js";
 import { withRetry, resolveRetryConfig } from "../core/retry.js";
 import type { RetryConfig } from "../core/retry.js";
 import {
@@ -45,132 +60,17 @@ import {
 } from "../input_request.js";
 import { runAbortableProcess } from "../abortable_process.js";
 
-export type WorkflowFile = {
-	name?: string;
-	description?: string;
-	args?: Record<string, { default?: unknown; description?: string }>;
-	env?: Record<string, string>;
-	cwd?: string;
-	steps: WorkflowStep[];
-	cost_limit?: CostLimit;
-};
-
-export type ParallelBranch = {
-	id: string;
-	run?: string;
-	command?: string;
-	pipeline?: string;
-	env?: Record<string, string>;
-	cwd?: string;
-	stdin?: unknown;
-};
-
-export type ParallelConfig = {
-	wait?: "all" | "any";
-	timeout_ms?: number;
-	branches: ParallelBranch[];
-};
-
-export type WorkflowStep = {
-	id: string;
-	command?: string;
-	run?: string;
-	pipeline?: string;
-	workflow?: string;
-	workflow_args?: Record<string, unknown>;
-	env?: Record<string, string>;
-	cwd?: string;
-	stdin?: unknown;
-	approval?: WorkflowApproval;
-	input?: WorkflowInputRequest;
-	condition?: unknown;
-	when?: unknown;
-	parallel?: ParallelConfig;
-	for_each?: string;
-	item_var?: string;
-	index_var?: string;
-	batch_size?: number;
-	pause_ms?: number;
-	steps?: WorkflowStep[];
-	timeout_ms?: number;
-	on_error?: "stop" | "continue" | "skip_rest";
-	retry?: {
-		max?: number;
-		backoff?: "fixed" | "exponential";
-		delay_ms?: number;
-		max_delay_ms?: number;
-		jitter?: boolean;
-	};
-};
-
-export type WorkflowApproval =
-	| boolean
-	| "required"
-	| string
-	| {
-			prompt?: string;
-			items?: unknown[];
-			preview?: string;
-			initiated_by?: string;
-			initiatedBy?: string;
-			required_approver?: string;
-			requiredApprover?: string;
-			require_different_approver?: boolean;
-			requireDifferentApprover?: boolean;
-	  };
-
-type WorkflowApprovalIdentity = {
-	initiatedBy?: string;
-	requiredApprover?: string;
-	requireDifferentApprover?: boolean;
-};
-
-export type WorkflowInputRequest = {
-	prompt: string;
-	responseSchema: unknown;
-	defaults?: unknown;
-};
-
-export type WorkflowStepResult = {
-	id: string;
-	stdout?: string;
-	json?: unknown;
-	approved?: boolean;
-	initiatedBy?: string;
-	approvedBy?: string;
-	subject?: unknown;
-	response?: unknown;
-	skipped?: boolean;
-	error?: boolean;
-	errorMessage?: string;
-};
-
-export type WorkflowRunResult = {
-	status: "ok" | "needs_approval" | "needs_input" | "cancelled";
-	output: unknown[];
-	requiresApproval?: {
-		type: "approval_request";
-		prompt: string;
-		items: unknown[];
-		preview?: string;
-		initiatedBy?: string;
-		requiredApprover?: string;
-		requireDifferentApprover?: boolean;
-		resumeToken?: string;
-		approvalId?: string;
-	};
-	requiresInput?: {
-		type: "input_request";
-		prompt: string;
-		responseSchema: unknown;
-		defaults?: unknown;
-		subject?: unknown;
-		resumeToken?: string;
-	};
-	_meta?: {
-		cost?: CostSummary;
-	};
-};
+export { loadWorkflowFile } from "./load.js";
+export type {
+	WorkflowFile,
+	ParallelBranch,
+	ParallelConfig,
+	WorkflowStep,
+	WorkflowApproval,
+	WorkflowInputRequest,
+	WorkflowStepResult,
+	WorkflowRunResult,
+} from "./types.js";
 
 type RunContext = {
 	stdin: NodeJS.ReadableStream;
@@ -316,422 +216,6 @@ class WorkflowPipelineInputSuspension extends Error {
 		this.request = request;
 		this.pipelineInput = pipelineInput;
 	}
-}
-
-export async function loadWorkflowFile(filePath: string): Promise<WorkflowFile> {
-	const text = await fsp.readFile(filePath, "utf8");
-	const ext = path.extname(filePath).toLowerCase();
-	const parsed = ext === ".json" ? JSON.parse(text) : parseYaml(text);
-
-	if (!parsed || typeof parsed !== "object") {
-		throw new Error("Workflow file must be a JSON/YAML object");
-	}
-
-	const steps = (parsed as WorkflowFile).steps;
-	if (!Array.isArray(steps) || steps.length === 0) {
-		throw new Error("Workflow file requires a non-empty steps array");
-	}
-
-	const costLimit = (parsed as WorkflowFile).cost_limit;
-	if (costLimit !== undefined) {
-		if (!costLimit || typeof costLimit !== "object" || Array.isArray(costLimit)) {
-			throw new Error("Workflow cost_limit must be an object");
-		}
-		if (!Number.isFinite(Number(costLimit.max_usd)) || Number(costLimit.max_usd) < 0) {
-			throw new Error("Workflow cost_limit.max_usd must be a non-negative number");
-		}
-		if (
-			costLimit.action !== undefined &&
-			costLimit.action !== "warn" &&
-			costLimit.action !== "stop"
-		) {
-			throw new Error('Workflow cost_limit.action must be "warn" or "stop"');
-		}
-	}
-
-	const seen = new Set<string>();
-	for (const step of steps) {
-		if (!step || typeof step !== "object") {
-			throw new Error("Workflow step must be an object");
-		}
-		if (!step.id || typeof step.id !== "string") {
-			throw new Error("Workflow step requires an id");
-		}
-		if (step.workflow !== undefined && typeof step.workflow !== "string") {
-			throw new Error(`Workflow step ${step.id} workflow must be a string (file path)`);
-		}
-		if (typeof step.workflow === "string" && !step.workflow.trim()) {
-			throw new Error(`Workflow step ${step.id} workflow path cannot be blank`);
-		}
-		if (step.workflow_args !== undefined) {
-			if (
-				!step.workflow_args ||
-				typeof step.workflow_args !== "object" ||
-				Array.isArray(step.workflow_args)
-			) {
-				throw new Error(`Workflow step ${step.id} workflow_args must be a plain object`);
-			}
-		}
-		if (
-			step.parallel !== undefined &&
-			(!step.parallel || typeof step.parallel !== "object" || Array.isArray(step.parallel))
-		) {
-			throw new Error(`Workflow step ${step.id} parallel must be an object`);
-		}
-		const isParallel = Boolean(
-			step.parallel && typeof step.parallel === "object" && !Array.isArray(step.parallel),
-		);
-		if (isParallel) {
-			const parallel = step.parallel as ParallelConfig;
-			if (!Array.isArray(parallel.branches) || parallel.branches.length === 0) {
-				throw new Error(`Workflow step ${step.id} parallel requires a non-empty branches array`);
-			}
-			if (parallel.wait !== undefined && parallel.wait !== "all" && parallel.wait !== "any") {
-				throw new Error(`Workflow step ${step.id} parallel wait must be "all" or "any"`);
-			}
-			if (
-				parallel.timeout_ms !== undefined &&
-				(typeof parallel.timeout_ms !== "number" ||
-					!Number.isFinite(parallel.timeout_ms) ||
-					!Number.isInteger(parallel.timeout_ms) ||
-					parallel.timeout_ms < 1 ||
-					parallel.timeout_ms > 2_147_483_647)
-			) {
-				throw new Error(
-					`Workflow step ${step.id} parallel timeout_ms must be a positive integer between 1 and 2147483647`,
-				);
-			}
-
-			const branchIds = new Set<string>();
-			for (const branch of parallel.branches) {
-				if (!branch || typeof branch !== "object") {
-					throw new Error(`Workflow step ${step.id} parallel branches must be objects`);
-				}
-				if (!branch.id || typeof branch.id !== "string") {
-					throw new Error(`Workflow step ${step.id} parallel branch requires an id`);
-				}
-				if (branch.id === step.id) {
-					throw new Error(`Workflow step ${step.id} parallel branch id cannot match the step id`);
-				}
-				if (branchIds.has(branch.id)) {
-					throw new Error(`Workflow step ${step.id} duplicate parallel branch id: ${branch.id}`);
-				}
-				if (seen.has(branch.id)) {
-					throw new Error(`Duplicate workflow id across steps/parallel branches: ${branch.id}`);
-				}
-				branchIds.add(branch.id);
-				const branchShell = typeof branch.run === "string" ? branch.run : branch.command;
-				const branchPipeline = typeof branch.pipeline === "string" ? branch.pipeline : undefined;
-				const branchExecCount = Number(Boolean(branchShell)) + Number(Boolean(branchPipeline));
-				if (branchExecCount === 0) {
-					throw new Error(
-						`Workflow step ${step.id} parallel branch ${branch.id} requires run, command, or pipeline`,
-					);
-				}
-				if (branchExecCount > 1) {
-					throw new Error(
-						`Workflow step ${step.id} parallel branch ${branch.id} can only define one of run, command, or pipeline`,
-					);
-				}
-				if (branch.run !== undefined && typeof branch.run !== "string") {
-					throw new Error(
-						`Workflow step ${step.id} parallel branch ${branch.id} run must be a string`,
-					);
-				}
-				if (branch.command !== undefined && typeof branch.command !== "string") {
-					throw new Error(
-						`Workflow step ${step.id} parallel branch ${branch.id} command must be a string`,
-					);
-				}
-				if (branch.pipeline !== undefined && typeof branch.pipeline !== "string") {
-					throw new Error(
-						`Workflow step ${step.id} parallel branch ${branch.id} pipeline must be a string`,
-					);
-				}
-			}
-		}
-		if (step.for_each !== undefined && typeof step.for_each !== "string") {
-			throw new Error(
-				`Workflow step ${step.id} for_each must be a string (step reference expression)`,
-			);
-		}
-		const isForEach = typeof step.for_each === "string";
-		if (isForEach) {
-			if (!Array.isArray(step.steps) || step.steps.length === 0) {
-				throw new Error(`Workflow step ${step.id} for_each requires a non-empty steps array`);
-			}
-			if (
-				step.batch_size !== undefined &&
-				(typeof step.batch_size !== "number" ||
-					!Number.isInteger(step.batch_size) ||
-					step.batch_size < 1)
-			) {
-				throw new Error(`Workflow step ${step.id} batch_size must be a positive integer`);
-			}
-			if (
-				step.pause_ms !== undefined &&
-				(typeof step.pause_ms !== "number" || !Number.isFinite(step.pause_ms) || step.pause_ms < 0)
-			) {
-				throw new Error(`Workflow step ${step.id} pause_ms must be a finite non-negative number`);
-			}
-			if (isApprovalStep(step.approval)) {
-				throw new Error(
-					`Workflow step ${step.id} for_each steps cannot define approval (use a separate step after the loop)`,
-				);
-			}
-			if (isInputStep(step.input)) {
-				throw new Error(
-					`Workflow step ${step.id} for_each steps cannot define input (use a separate step after the loop)`,
-				);
-			}
-			if (step.stdin !== undefined && step.stdin !== null) {
-				throw new Error(
-					`Workflow step ${step.id} for_each steps cannot define stdin (loop input comes from the for_each expression)`,
-				);
-			}
-			const loopShell = typeof step.run === "string" ? step.run : step.command;
-			const loopPipeline = typeof step.pipeline === "string" ? step.pipeline : undefined;
-			if (loopShell || loopPipeline || step.workflow || step.parallel) {
-				throw new Error(
-					`Workflow step ${step.id} for_each cannot also define run, command, pipeline, workflow, or parallel`,
-				);
-			}
-			if (step.item_var !== undefined && typeof step.item_var !== "string") {
-				throw new Error(`Workflow step ${step.id} item_var must be a string`);
-			}
-			if (step.index_var !== undefined && typeof step.index_var !== "string") {
-				throw new Error(`Workflow step ${step.id} index_var must be a string`);
-			}
-			const loopItemVar = step.item_var ?? "item";
-			const loopIndexVar = step.index_var ?? "index";
-			if (loopItemVar === loopIndexVar) {
-				throw new Error(`Workflow step ${step.id} item_var and index_var cannot be the same`);
-			}
-			const subStepIds = new Set<string>();
-			for (const sub of step.steps) {
-				if (!sub || typeof sub !== "object" || !sub.id || typeof sub.id !== "string") {
-					throw new Error(`Workflow step ${step.id} for_each sub-step requires an id`);
-				}
-				if (sub.id === loopItemVar || sub.id === loopIndexVar) {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step id '${sub.id}' conflicts with loop variable`,
-					);
-				}
-				if (subStepIds.has(sub.id)) {
-					throw new Error(`Workflow step ${step.id} duplicate for_each sub-step id: ${sub.id}`);
-				}
-				subStepIds.add(sub.id);
-				if (isApprovalStep(sub.approval) || isInputStep(sub.input)) {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-steps cannot contain approval or input steps`,
-					);
-				}
-				if (sub.run !== undefined && typeof sub.run !== "string") {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step ${sub.id} run must be a string`,
-					);
-				}
-				if (sub.command !== undefined && typeof sub.command !== "string") {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step ${sub.id} command must be a string`,
-					);
-				}
-				if (sub.pipeline !== undefined && typeof sub.pipeline !== "string") {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step ${sub.id} pipeline must be a string`,
-					);
-				}
-				if (sub.workflow || sub.parallel || sub.for_each) {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step ${sub.id} cannot define workflow, parallel, or for_each`,
-					);
-				}
-				const subShell =
-					typeof sub.run === "string" && sub.run.trim()
-						? sub.run
-						: typeof sub.command === "string" && sub.command.trim()
-							? sub.command
-							: undefined;
-				const subPipeline =
-					typeof sub.pipeline === "string" && sub.pipeline.trim() ? sub.pipeline : undefined;
-				if (!subShell && !subPipeline) {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step ${sub.id} requires run, command, or pipeline`,
-					);
-				}
-				if (Number(Boolean(subShell)) + Number(Boolean(subPipeline)) > 1) {
-					throw new Error(
-						`Workflow step ${step.id} for_each sub-step ${sub.id} can only define one of run, command, or pipeline`,
-					);
-				}
-			}
-		}
-		const shellCommand = typeof step.run === "string" ? step.run : step.command;
-		const pipeline = typeof step.pipeline === "string" ? step.pipeline : undefined;
-		const workflowRef =
-			typeof step.workflow === "string" && step.workflow.trim() ? step.workflow : undefined;
-		const executionCount =
-			Number(Boolean(shellCommand)) +
-			Number(Boolean(pipeline)) +
-			Number(Boolean(workflowRef)) +
-			Number(isParallel) +
-			Number(isForEach);
-		if (executionCount === 0 && !isApprovalStep(step.approval) && !isInputStep(step.input)) {
-			throw new Error(
-				`Workflow step ${step.id} requires run, command, pipeline, workflow, parallel, for_each, approval, or input`,
-			);
-		}
-		if (executionCount > 1) {
-			throw new Error(
-				`Workflow step ${step.id} can only define one of run, command, pipeline, workflow, parallel, or for_each`,
-			);
-		}
-		if (executionCount > 0 && isInputStep(step.input)) {
-			throw new Error(
-				`Workflow step ${step.id} input steps cannot define run, command, pipeline, workflow, parallel, or for_each`,
-			);
-		}
-		if (isApprovalStep(step.approval) && isInputStep(step.input)) {
-			throw new Error(`Workflow step ${step.id} cannot define both approval and input`);
-		}
-		if (step.run !== undefined && typeof step.run !== "string") {
-			throw new Error(`Workflow step ${step.id} run must be a string`);
-		}
-		if (step.command !== undefined && typeof step.command !== "string") {
-			throw new Error(`Workflow step ${step.id} command must be a string`);
-		}
-		if (step.pipeline !== undefined && typeof step.pipeline !== "string") {
-			throw new Error(`Workflow step ${step.id} pipeline must be a string`);
-		}
-		if (step.input !== undefined && !isInputStep(step.input)) {
-			throw new Error(`Workflow step ${step.id} input must be an object`);
-		}
-		if (step.input && typeof step.input.prompt !== "string") {
-			throw new Error(`Workflow step ${step.id} input.prompt must be a string`);
-		}
-		if (
-			step.input &&
-			(step.input.responseSchema === undefined || typeof step.input.responseSchema !== "object")
-		) {
-			throw new Error(`Workflow step ${step.id} input.responseSchema must be an object`);
-		}
-		if (step.input) {
-			try {
-				compileCached(step.input.responseSchema as any);
-			} catch (err: any) {
-				throw new Error(
-					`Workflow step ${step.id} input.responseSchema is invalid: ${err?.message ?? String(err)}`,
-				);
-			}
-		}
-		if (step.approval && typeof step.approval === "object" && !Array.isArray(step.approval)) {
-			const approval = step.approval as Record<string, unknown>;
-			if (approval.initiated_by !== undefined && typeof approval.initiated_by !== "string") {
-				throw new Error(`Workflow step ${step.id} approval.initiated_by must be a string`);
-			}
-			if (approval.initiatedBy !== undefined && typeof approval.initiatedBy !== "string") {
-				throw new Error(`Workflow step ${step.id} approval.initiatedBy must be a string`);
-			}
-			if (
-				approval.required_approver !== undefined &&
-				typeof approval.required_approver !== "string"
-			) {
-				throw new Error(`Workflow step ${step.id} approval.required_approver must be a string`);
-			}
-			if (
-				approval.requiredApprover !== undefined &&
-				typeof approval.requiredApprover !== "string"
-			) {
-				throw new Error(`Workflow step ${step.id} approval.requiredApprover must be a string`);
-			}
-			if (
-				approval.require_different_approver !== undefined &&
-				typeof approval.require_different_approver !== "boolean"
-			) {
-				throw new Error(
-					`Workflow step ${step.id} approval.require_different_approver must be a boolean`,
-				);
-			}
-			if (
-				approval.requireDifferentApprover !== undefined &&
-				typeof approval.requireDifferentApprover !== "boolean"
-			) {
-				throw new Error(
-					`Workflow step ${step.id} approval.requireDifferentApprover must be a boolean`,
-				);
-			}
-		}
-		if (
-			step.timeout_ms !== undefined &&
-			(typeof step.timeout_ms !== "number" ||
-				!Number.isFinite(step.timeout_ms) ||
-				!Number.isInteger(step.timeout_ms) ||
-				step.timeout_ms < 1 ||
-				step.timeout_ms > 2_147_483_647)
-		) {
-			throw new Error(
-				`Workflow step ${step.id} timeout_ms must be a positive integer between 1 and 2147483647`,
-			);
-		}
-		if (
-			step.on_error !== undefined &&
-			step.on_error !== "stop" &&
-			step.on_error !== "continue" &&
-			step.on_error !== "skip_rest"
-		) {
-			throw new Error(
-				`Workflow step ${step.id} on_error must be "stop", "continue", or "skip_rest"`,
-			);
-		}
-		if (step.retry !== undefined) {
-			if (!step.retry || typeof step.retry !== "object" || Array.isArray(step.retry)) {
-				throw new Error(`Workflow step ${step.id} retry must be an object`);
-			}
-			const r = step.retry;
-			if (
-				r.max !== undefined &&
-				(typeof r.max !== "number" || !Number.isInteger(r.max) || r.max < 1)
-			) {
-				throw new Error(`Workflow step ${step.id} retry.max must be a positive integer`);
-			}
-			if (r.backoff !== undefined && r.backoff !== "fixed" && r.backoff !== "exponential") {
-				throw new Error(`Workflow step ${step.id} retry.backoff must be "fixed" or "exponential"`);
-			}
-			if (
-				r.delay_ms !== undefined &&
-				(typeof r.delay_ms !== "number" || !Number.isFinite(r.delay_ms) || r.delay_ms < 0)
-			) {
-				throw new Error(
-					`Workflow step ${step.id} retry.delay_ms must be a finite non-negative number`,
-				);
-			}
-			if (
-				r.max_delay_ms !== undefined &&
-				(typeof r.max_delay_ms !== "number" ||
-					!Number.isFinite(r.max_delay_ms) ||
-					r.max_delay_ms < 0)
-			) {
-				throw new Error(
-					`Workflow step ${step.id} retry.max_delay_ms must be a finite non-negative number`,
-				);
-			}
-			if (r.jitter !== undefined && typeof r.jitter !== "boolean") {
-				throw new Error(`Workflow step ${step.id} retry.jitter must be a boolean`);
-			}
-		}
-		if (seen.has(step.id)) {
-			throw new Error(`Duplicate workflow step id: ${step.id}`);
-		}
-		if (isParallel) {
-			const parallel = step.parallel as ParallelConfig;
-			for (const branch of parallel.branches) {
-				seen.add(branch.id);
-			}
-		}
-		seen.add(step.id);
-	}
-
-	return parsed as WorkflowFile;
 }
 
 export function resolveWorkflowArgs(
@@ -1326,7 +810,7 @@ export async function runWorkflowFile({
 						const runBranch = async (
 							branch: ParallelBranch,
 						): Promise<{ branchId: string; result: WorkflowStepResult }> => {
-							const mergedBranchEnv = { ...(step.env ?? {}), ...(branch.env ?? {}) };
+							const mergedBranchEnv = { ...step.env, ...branch.env };
 							const branchEnv = mergeEnv(
 								ctx.env,
 								workflow.env,
@@ -2548,15 +2032,6 @@ function resolveShellStdin(
 	return encodeShellInput(value);
 }
 
-function resolveTemplate(
-	input: string,
-	args: Record<string, unknown>,
-	results: Record<string, WorkflowStepResult>,
-) {
-	const withArgs = resolveArgsTemplate(input, args);
-	return resolveStepRefs(withArgs, results);
-}
-
 function resolveWorkflowStepArgs(
 	workflowArgs: Record<string, unknown> | undefined,
 	parentArgs: Record<string, unknown>,
@@ -2572,71 +2047,6 @@ function resolveWorkflowStepArgs(
 		}
 	}
 	return resolved;
-}
-
-function resolveArgsTemplate(input: string, args: Record<string, unknown>) {
-	return input.replace(/\$\{([A-Za-z0-9_-]+)\}/g, (match, key) => {
-		if (Object.hasOwn(args, key)) return String(args[key]);
-		return match;
-	});
-}
-
-function resolveStepRefs(input: string, results: Record<string, WorkflowStepResult>) {
-	return input.replace(
-		/\$([A-Za-z0-9_-]+)\.([A-Za-z0-9_]+(?:\.[A-Za-z0-9_]+)*)/g,
-		(match, id, pathValue) => {
-			if (!(id in results)) {
-				return match;
-			}
-			const refValue = getStepRefValue({ id, path: pathValue }, results, false);
-			if (refValue === undefined) {
-				if (pathValue === "approved" || pathValue === "skipped") return "false";
-				return "";
-			}
-			return renderTemplateValue(refValue);
-		},
-	);
-}
-
-function parseStepRef(value: string) {
-	const match = value.match(/^\$([A-Za-z0-9_-]+)\.([A-Za-z0-9_]+(?:\.[A-Za-z0-9_]+)*)$/);
-	if (!match) return null;
-	return { id: match[1], path: match[2] };
-}
-
-function getStepRefValue(
-	ref: { id: string; path: string },
-	results: Record<string, WorkflowStepResult>,
-	strict: boolean,
-) {
-	const step = results[ref.id];
-	if (!step) {
-		if (strict) throw new Error(`Unknown step reference: ${ref.id}.${ref.path}`);
-		return undefined;
-	}
-	return getValueByPath(step, ref.path);
-}
-
-function evaluateCondition(condition: unknown, results: Record<string, WorkflowStepResult>) {
-	if (condition === undefined || condition === null) return true;
-	if (typeof condition === "boolean") return condition;
-	if (typeof condition !== "string") throw new Error("Unsupported condition type");
-
-	const trimmed = condition.trim();
-	if (trimmed === "true") return true;
-	if (trimmed === "false") return false;
-	return evaluateConditionExpression(trimmed, results);
-}
-
-function isApprovalStep(approval: WorkflowStep["approval"]) {
-	if (approval === true) return true;
-	if (typeof approval === "string" && approval.trim().length > 0) return true;
-	if (approval && typeof approval === "object" && !Array.isArray(approval)) return true;
-	return false;
-}
-
-function isInputStep(input: WorkflowStep["input"]) {
-	return Boolean(input && typeof input === "object" && !Array.isArray(input));
 }
 
 function extractApprovalRequest(
@@ -3234,319 +2644,6 @@ function estimateSerializedBytes(value: unknown) {
 	} catch {
 		return Number.POSITIVE_INFINITY;
 	}
-}
-
-function renderTemplateValue(value: unknown) {
-	if (value === undefined || value === null) return "";
-	if (typeof value === "string") return value;
-	if (typeof value === "number" || typeof value === "boolean" || typeof value === "bigint") {
-		return String(value);
-	}
-	try {
-		return JSON.stringify(value);
-	} catch {
-		return String(value);
-	}
-}
-
-function getValueByPath(value: unknown, pathValue: string) {
-	const fields = pathValue.split(".");
-	let current: unknown = value;
-	for (const field of fields) {
-		if (current === null || current === undefined) return undefined;
-		if (Array.isArray(current)) {
-			const idx = Number(field);
-			if (!Number.isInteger(idx) || idx < 0 || idx >= current.length) return undefined;
-			current = current[idx];
-			continue;
-		}
-		if (typeof current !== "object") return undefined;
-		current = (current as Record<string, unknown>)[field];
-	}
-	return current;
-}
-
-type ConditionToken =
-	| {
-			type: "lparen" | "rparen" | "and" | "or" | "eq" | "neq" | "lt" | "lte" | "gt" | "gte" | "not";
-	  }
-	| { type: "step_ref"; value: { id: string; path: string } }
-	| { type: "string" | "number" | "boolean" | "null" | "identifier"; value: unknown };
-
-function evaluateConditionExpression(
-	expression: string,
-	results: Record<string, WorkflowStepResult>,
-) {
-	const tokens = tokenizeCondition(expression);
-	if (tokens.length === 0) {
-		throw new Error(`Unsupported condition: ${expression}`);
-	}
-	let index = 0;
-
-	function parseOr(): unknown {
-		let left = parseAnd();
-		while (match("or")) {
-			const right = parseAnd();
-			left = Boolean(left) || Boolean(right);
-		}
-		return left;
-	}
-
-	function parseAnd(): unknown {
-		let left = parseEquality();
-		while (match("and")) {
-			const right = parseEquality();
-			left = Boolean(left) && Boolean(right);
-		}
-		return left;
-	}
-
-	function parseEquality(): unknown {
-		const left = parseUnary(false);
-		if (match("eq")) {
-			return compareConditionValues(left, parseUnary(true));
-		}
-		if (match("neq")) {
-			return !compareConditionValues(left, parseUnary(true));
-		}
-		if (match("lt")) {
-			return numericCompare(left, parseUnary(true), (a, b) => a < b);
-		}
-		if (match("lte")) {
-			return numericCompare(left, parseUnary(true), (a, b) => a <= b);
-		}
-		if (match("gt")) {
-			return numericCompare(left, parseUnary(true), (a, b) => a > b);
-		}
-		if (match("gte")) {
-			return numericCompare(left, parseUnary(true), (a, b) => a >= b);
-		}
-		return left;
-	}
-
-	function parseUnary(allowBareIdentifier: boolean): unknown {
-		if (match("not")) {
-			return !Boolean(parseUnary(allowBareIdentifier));
-		}
-		return parsePrimary(allowBareIdentifier);
-	}
-
-	function parsePrimary(allowBareIdentifier: boolean): unknown {
-		const token = tokens[index];
-		if (!token) {
-			throw new Error(`Unsupported condition: ${expression}`);
-		}
-		index += 1;
-
-		if (token.type === "lparen") {
-			const value = parseOr();
-			expect("rparen");
-			return value;
-		}
-		if (token.type === "step_ref") {
-			return getStepRefValue(token.value, results, true);
-		}
-		if (
-			token.type === "string" ||
-			token.type === "number" ||
-			token.type === "boolean" ||
-			token.type === "null"
-		) {
-			return token.value;
-		}
-		if (token.type === "identifier" && allowBareIdentifier) {
-			return token.value;
-		}
-		throw new Error(`Unsupported condition: ${expression}`);
-	}
-
-	function match(type: ConditionToken["type"]) {
-		if (tokens[index]?.type !== type) return false;
-		index += 1;
-		return true;
-	}
-
-	function expect(type: ConditionToken["type"]) {
-		if (!match(type)) {
-			throw new Error(`Unsupported condition: ${expression}`);
-		}
-	}
-
-	const value = parseOr();
-	if (index !== tokens.length) {
-		throw new Error(`Unsupported condition: ${expression}`);
-	}
-	return Boolean(value);
-}
-
-function compareConditionValues(left: unknown, right: unknown) {
-	if (
-		Array.isArray(left) ||
-		Array.isArray(right) ||
-		isPlainConditionObject(left) ||
-		isPlainConditionObject(right)
-	) {
-		return isDeepStrictEqual(left, right);
-	}
-	return Object.is(left, right);
-}
-
-function isStrictlyNumeric(value: unknown): boolean {
-	if (typeof value === "number") return !Number.isNaN(value);
-	if (typeof value === "string") return value.trim() !== "" && !Number.isNaN(Number(value));
-	return false;
-}
-
-function numericCompare(
-	left: unknown,
-	right: unknown,
-	cmp: (a: number, b: number) => boolean,
-): boolean {
-	if (!isStrictlyNumeric(left) || !isStrictlyNumeric(right)) return false;
-	return cmp(Number(left), Number(right));
-}
-
-function isPlainConditionObject(value: unknown): value is Record<string, unknown> {
-	return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
-function tokenizeCondition(expression: string): ConditionToken[] {
-	const tokens: ConditionToken[] = [];
-	let index = 0;
-
-	while (index < expression.length) {
-		const ch = expression[index];
-		if (/\s/.test(ch)) {
-			index += 1;
-			continue;
-		}
-		if (ch === "(") {
-			tokens.push({ type: "lparen" });
-			index += 1;
-			continue;
-		}
-		if (ch === ")") {
-			tokens.push({ type: "rparen" });
-			index += 1;
-			continue;
-		}
-		if (expression.startsWith("&&", index)) {
-			tokens.push({ type: "and" });
-			index += 2;
-			continue;
-		}
-		if (expression.startsWith("||", index)) {
-			tokens.push({ type: "or" });
-			index += 2;
-			continue;
-		}
-		if (expression.startsWith("==", index)) {
-			tokens.push({ type: "eq" });
-			index += 2;
-			continue;
-		}
-		if (expression.startsWith("!=", index)) {
-			tokens.push({ type: "neq" });
-			index += 2;
-			continue;
-		}
-		if (expression.startsWith("<=", index)) {
-			tokens.push({ type: "lte" });
-			index += 2;
-			continue;
-		}
-		if (expression.startsWith(">=", index)) {
-			tokens.push({ type: "gte" });
-			index += 2;
-			continue;
-		}
-		if (ch === "<") {
-			tokens.push({ type: "lt" });
-			index += 1;
-			continue;
-		}
-		if (ch === ">") {
-			tokens.push({ type: "gt" });
-			index += 1;
-			continue;
-		}
-		if (ch === "!") {
-			tokens.push({ type: "not" });
-			index += 1;
-			continue;
-		}
-		if (ch === "$") {
-			const matched = matchConditionStepRef(expression, index);
-			if (!matched) {
-				throw new Error(`Unsupported condition: ${expression}`);
-			}
-			tokens.push({ type: "step_ref", value: matched.ref });
-			index = matched.nextIndex;
-			continue;
-		}
-		if (ch === '"' || ch === "'") {
-			const parsed = parseQuotedConditionString(expression, index, ch);
-			tokens.push({ type: "string", value: parsed.value });
-			index = parsed.nextIndex;
-			continue;
-		}
-		const numberMatch = expression.slice(index).match(/^-?\d+(?:\.\d+)?/);
-		if (numberMatch) {
-			tokens.push({ type: "number", value: Number(numberMatch[0]) });
-			index += numberMatch[0].length;
-			continue;
-		}
-		const identMatch = expression.slice(index).match(/^[A-Za-z_][A-Za-z0-9_-]*/);
-		if (identMatch) {
-			const raw = identMatch[0];
-			if (raw === "true") {
-				tokens.push({ type: "boolean", value: true });
-			} else if (raw === "false") {
-				tokens.push({ type: "boolean", value: false });
-			} else if (raw === "null") {
-				tokens.push({ type: "null", value: null });
-			} else {
-				tokens.push({ type: "identifier", value: raw });
-			}
-			index += raw.length;
-			continue;
-		}
-		throw new Error(`Unsupported condition: ${expression}`);
-	}
-
-	return tokens;
-}
-
-function matchConditionStepRef(expression: string, startIndex: number) {
-	const match = expression
-		.slice(startIndex)
-		.match(/^\$([A-Za-z0-9_-]+)\.([A-Za-z0-9_]+(?:\.[A-Za-z0-9_]+)*)/);
-	if (!match) return null;
-	return {
-		ref: { id: match[1], path: match[2] },
-		nextIndex: startIndex + match[0].length,
-	};
-}
-
-function parseQuotedConditionString(expression: string, startIndex: number, quoteChar: '"' | "'") {
-	let value = "";
-	let index = startIndex + 1;
-	while (index < expression.length) {
-		const ch = expression[index];
-		if (ch === "\\") {
-			const next = expression[index + 1];
-			if (next === undefined) break;
-			value += next;
-			index += 2;
-			continue;
-		}
-		if (ch === quoteChar) {
-			return { value, nextIndex: index + 1 };
-		}
-		value += ch;
-		index += 1;
-	}
-	throw new Error(`Unsupported condition: ${expression}`);
 }
 
 async function runShellCommand({
